Modif int en long et config bdd h2

+++ src/main/java/fr/meya/witcher_toolkit_back/model/persistent/Campagne.java
@@ -1,11 +1,14 @@
 package fr.meya.witcher_toolkit_back.model.persistent;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.Setter;
-
-import java.util.List;
 
 /**
  * Campagne géré par les utilisateurs.
@@ -17,7 +20,7 @@ public class Campagne {
 
 	@Id
 	@GeneratedValue( strategy = GenerationType.IDENTITY )
-	private int idCampagne;
+	private long idCampagne;
 
 	@NotBlank
 	private String nom;
